+++ labbook_BE/app/__init__.py
@@ -72,11 +72,20 @@ if config_envvar in os.environ:
     print("Loading local configuration from {}={}".format(config_envvar, os.environ[config_envvar]))
     app.config.from_envvar(config_envvar)
 
+    os.environ['LABBOOK_KEY_DIR']    = Constants.cst_key
+    os.environ['LABBOOK_STATUS_DIR'] = Constants.cst_io
+    os.environ['LABBOOK_LOG_DIR']    = Constants.cst_log
+    os.environ['LABBOOK_USER']       = Constants.cst_script_user
+
     # Put in os.environ DB variables
-    os.environ['DB_USER'] = app.config['DB_USER']
-    os.environ['DB_PWD']  = app.config['DB_PWD']
-    os.environ['DB_HOST'] = app.config['DB_HOST']
-    os.environ['DB_NAME'] = app.config['DB_NAME']
+    os.environ['LABBOOK_DB_USER'] = app.config['DB_USER']
+    os.environ['LABBOOK_DB_PWD']  = app.config['DB_PWD']
+    os.environ['LABBOOK_DB_HOST'] = app.config['DB_HOST']
+    os.environ['LABBOOK_DB_NAME'] = app.config['DB_NAME']
+
+    # ONLY FOR TESTS 
+    #os.environ['LABBOOK_TEST_OK']= Constants.cst_io_keyexist + ',' + Constants.cst_io_listmedia + ',' + Constants.cst_io_progbackup + ',' + Constants.cst_io_initmedia + ',' + Constants.cst_io_backup + ',' + Constants.cst_io_listarchive + ',' + Constants.cst_io_restore
+    # os.environ['LABBOOK_TEST_KO']= Constants.cst_io_genkey + ',' + Constants.cst_io_initmedia
 else:
     print("No local configuration available: {} is undefined in the environment".format(config_envvar))
 
@@ -177,13 +186,18 @@ api.add_resource(ResultReset,         '/services/result/reset/<int:id_rec>')
 api.add_resource(ResultValid,         '/services/result/valid/<string:type_valid>/<int:id_rec>')
 api.add_resource(ResultValue,         '/services/result/list/value')
 api.add_resource(SettingAgeInterval,  '/services/setting/age/interval')
+api.add_resource(SettingBackup,       '/services/setting/backup')
 api.add_resource(SettingPref,         '/services/setting/pref/list', '/services/setting/pref/list/<int:id_owner>')
 api.add_resource(SettingRecNum,       '/services/setting/record/number')
 api.add_resource(SettingReport,       '/services/setting/report')
-api.add_resource(SettingBackup,       '/services/setting/backup')
 api.add_resource(SettingSticker,      '/services/setting/sticker', '/services/setting/sticker/<int:sts_ser>')
-api.add_resource(ScriptBackup,        '/services/setting/script/backup')
+api.add_resource(ScriptBackup,        '/services/setting/script/backup/<string:media>')
+api.add_resource(ScriptGenkey,        '/services/setting/script/genkey')
+api.add_resource(ScriptInitmedia,     '/services/setting/script/initmedia/<string:media>')
 api.add_resource(ScriptKeyexist,      '/services/setting/script/keyexist')
+api.add_resource(ScriptListarchive,   '/services/setting/script/listarchive/<string:media>')
+api.add_resource(ScriptListmedia,     '/services/setting/script/listmedia/<string:type>')
+api.add_resource(ScriptProgbackup,    '/services/setting/script/progbackup/<string:hour>')
 api.add_resource(ScriptRestore,       '/services/setting/script/restore')
 api.add_resource(StaffExport,         '/services/quality/staff/export')
 api.add_resource(StockProductDet,     '/services/quality/stock/product/det/<int:id_item>')